server/requirements.txt
openenv-core>=0.2.2
mujoco>=3.1.0
numpy>=1.24.0
hydra-core>=1.3.0
omegaconf>=2.3.0
pandas>=2.0.0
scipy>=1.11.0
pyyaml>=6.0
imageio>=2.34.0
tqdm>=4.66.0
ipdb>=0.13.13
